api: move manufactoring logic into the main scope
