+++ manufacturing.py
